docker/requirements.txt
Flask
pybase64
jsonpickle
pillow
numpy
opencv-python
